solve duplication of test_iraf_trace
